+++ src/Sentry/Extensibility/IBackgroundWorker.cs
@@ -20,7 +20,7 @@ namespace Sentry.Extensibility
         /// </summary>
         /// <param name="timeout">How long to wait for flush to finish.</param>
         /// <returns>A task to await for the flush operation.</returns>
-        Task FlushAsync(TimeSpan timeout);
+        ValueTask FlushAsync(TimeSpan timeout);
 
         /// <summary>
         /// Current count of items queued up.